Simplify Heroku deploiement (Details in commit description)
Package added which will simplify the deploiement on Heroku:
- `whitenoise`
- `dj-database-url`
- `python-dotenv` (replace the `django-environ` package)

The `django-environ` has been removed because he complexify the deploiement on Heroku because an exception is throw when the `.env' file does not exist (and the `.env' file is not in source control because of the sensitive data it may contain).

The `.env.example` has been updated and all `DB_` variables are replaced by the `DATABASE_URL`.

The README.md has been updated accordingly on the `.env` variables to fill when installing the project.

A `release` command has been added to the `Procfile` which will execute a shell script that will perform some actions (migrations, imports, etc...)