Drop python version 3.7 & 3.8 for travis builds
I decided to drop 2 of the 3 builds (3.7 and 3.8 versions). For a reason that I can't explain, builds started to fail randomly due to existing database sessions. I tried the project in each python versions, try to check what difference was made on each commits when the builds starts to fail. But nothing interesting. A weird thing is that the builds who works sucessfully before now fails too. I can't figure it out so, I'll only build one version by now.