Limit date equality accuracy in `review.html`
Before the equality was tested and both dates hold miliseconds but, when created a review, there's a slight difference in ms between `created_at` and `updated_at`.

So, in the template, when we were creating a fresh review, next to the creation date, the updated date was present too caused by this slight difference.

Now we check the equality with an accuracy of one second.